requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
blis==0.7.9
Brotli==1.0.9
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
en-core-web-lg==3.5.0
filelock==3.9.0
Flask==2.3.2
Flask-Cors==4.0.0
flask-siwadoc==0.2.1
frozenlist==1.3.3
fsspec==2023.6.0
gevent==22.10.2
geventhttpclient==2.0.9
greenlet==2.0.2
grpcio==1.37.1
grpcio-tools==1.37.1
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
MarkupSafe==2.1.3
mpmath==1.2.1
multidict==6.0.4
murmurhash==1.0.9
nebula3-database==0.3.2
nebula3-experiments==0.0.13
nebula3-experts==1.6.7
nebula3-vg-driver==0.0.2
networkx==3.0
nltk==3.5
numpy==1.24.4
openai==0.27.8
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
pathy==0.10.2
Pillow==9.5.0
preshed==3.0.8
protobuf==3.20.3
pydantic==1.9.0
PyJWT==2.7.0
pymilvus==1.1.1
python-arango==7.2.0
python-dateutil==2.8.2
python-rapidjson==1.10
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.29.0
requests-toolbelt==1.0.0
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools-scm==7.1.0
six==1.16.0
smart-open==6.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-wordnet==0.1.0
srsly==2.4.6
sympy==1.11.1
thinc==8.1.10
threadpoolctl==3.1.0
tokenizers==0.13.3
tomli==2.0.1
tqdm==4.65.0
transformers==4.30.2
tritonclient==2.34.0
typer==0.7.0
typing-extensions==4.5.0
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.16
wasabi==1.1.2
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.15.0
zope.event==5.0
zope.interface==6.0
torch==2.0.1
torchaudio==0.12.0+cu116
torchvision==0.15.2